Method abort_pre_commit() required to report BF abort success.
Changed abort_pre_commit() requirements so that more information
about BF abort success is propageted to the caller. The call is
now required to return:

* WSREP_TRX_MISSING if the provider could not perform BF abort
  if the victim_trx was not found by the provider
* WSREP_NO_ALLOWED if the provider declined to perform BF abort
  for any other reason

Additionally the abort_pre_commit() is now required to return
victim_seqno if the BF abort was succesful. This is needed
if the commit ordering management is done in the application
which uses wsrep provider.